+++ src/app/my-page/bookmarks/posts/_components/BookmarkPostsList.tsx
@@ -7,7 +7,7 @@
 'use client';
 
 import PaginationWrapper from '@/components/ui/PaginationWrapper';
-import { TransformedBookmarkItem } from '@/lib/functions/mypage/bookmarkFunctions';
+import { TransformedBookmarkItem } from '@/lib/functions/mypage/bookmark/bookmarkFunctions';
 import { BookOpen } from 'lucide-react';
 import { useRouter, useSearchParams } from 'next/navigation';
 import { useEffect, useMemo, useRef, useState } from 'react';
