+++ src/app/my-page/bookmarks/products/_components/ProductCard.tsx
@@ -5,7 +5,7 @@
 'use client';
 
 import { Button } from '@/components/ui/Button';
-import { deleteBookmark } from '@/lib/actions/mypage/bookmarkActions';
+import { deleteBookmark } from '@/lib/actions/mypage/bookmark/bookmarkActions';
 import { Trash2 } from 'lucide-react';
 import Image from 'next/image';
 import { useState } from 'react';
